Post a post to Wordpress

Post a post to Wordpress

wp_post(post_folder, wordpress_url, encoding = "UTF-8")

Arguments

post_folder

Path to folder where the post (index.Rmd with hugodown::md_document output format, index.md, images) lives

wordpress_url

URL to your website

encoding

Encoding argument to readLines

Value

URL to the post (invisibly)

Details

Several aspects of the post can be controlled by the YAML of index.Rmd.

---
title: "Title of the Post" # compulsory!
date: "2020-04-01T00:00:00" # compulsory!
author: "admin7891" # Either omit the field (the author will be
#  the authenticated user) or use an username
slug: "post-slug" # Important especially if the slug is in the URL
excerpt: "Here I summarize this fantastic post"
status: "publish" # One of: "publish", "future", "draft",
#  "pending", "private"; "publish" by default
output: hugodown::md_document # Don't change this
categories: # Categories, optional, will be created if
# the post is the first with the category
  - math
- Code and Stuff
tags: # Tags, optional, will be created if
# the post is the first with the tag
  - crul
  - mathjax
  - R packages
comment_status: closed # Either closed (default) or open
ping_status: closed # Either closed (default) or open
---

The package cannot handle private posts with password, only private posts that are visible to admins and editors only. You could create a private post, and then from the WordPress interface make it visible with password. Make it private again before trying to update the post with the R package.

Examples

if (FALSE) {
# this requires authentication!
post_folder <- system.file(file.path("post-example"), package = "goodpress")
tmp_post_folder <- file.path(tempdir(), "post")
dir.create(tmp_post_folder)
file.copy(dir(post_folder, full.names = TRUE), tmp_post_folder)
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_post(tmp_post_folder, wordpress_url)
file.remove(dir(tmp_post_folder, full.names = TRUE))
}